+++ ReactViewResources/Loader/ViewContext.ts
@@ -1,0 +1,4 @@
+﻿import * as React from "react";
+import { ViewMetadata } from "./ViewMetadata";
+
+export const ViewContext = React.createContext<ViewMetadata>(null!);
